Structurize app directory
